Make sure criticality_score between 0 and 1
Due to the negative weight, the numerator of score may be greater than the denominator, that means the score would be greater than 1. And the numerator may be also less than 0, that means final score would be less than 0.

So in this patch, we only summerize all positive weight, and set the negative criticality_score to 0 (least-critical).