Add a command for enumerating GitHub repos (written in Go)
Takes the algorithms suggested in #33 and combines them to allow approx
repos with at least 10 stars and higher to be enumerated.

This approach *only* collects project URLs.

There are other sources of data that can be used such as GH Archive,
GHTorrent, etc to compliment this list.

All the project URLs found can eventually be de-duped, checked for
existance (e.g. 404s), canonicalized, and then used for signal collection.