Update k8s config to remove controller and pubsub for local version.
This change replaces the GCP PubSub with the local, single worker
version of collect_signals.

The init_collect_signals script is used to replace the work of the
Scorecard controller.

I have tested the staging side of these changes and will verify the
production side later next week when this change and staging are
verifdied to be working correctly.

Signed-off-by: Caleb Brown <calebbrown@google.com>